add ranking module and move TopN
