+++ doc/basic.rst
@@ -27,20 +27,6 @@ The :py:class:`Random` algorithm implements random-item recommendation.
     :show-inheritance:
 
 
-Top-N Recommender
------------------
-
-The :py:class:`TopN` class implements a standard top-*N* recommender that wraps a
-:py:class:`.Predictor` and :py:class:`.CandidateSelector` and returns the top *N*
-candidate items by predicted rating.  It is the type of recommender returned by
-:py:meth:`.Recommender.adapt` if the provided algorithm is not a recommender.
-
-.. module:: lenskit.algorithms.basic
-.. autoclass:: TopN
-    :members:
-    :show-inheritance:
-
-
 Unrated Item Candidate Selector
 -------------------------------
